Support iOS9, bug fixes, header cleanup and reformat
Add support for iOS 9 and 9.1
Fixed small memory leak when using a shadow border
Made tree/treetables use dynamic icon
Fixed some copyright headers and reformatted source
Removed support for iOS6
Upgraded JSONParser to support a watcher interface to support pull
parsing
Made JSON object use SNumber when returning number objects
Added stream handler support to URL class on iOS
Fixed fling support on google glass
Ma

+++ source/rare/apple/com/appnativa/rare/platform/apple/ui/view/PopupWindow.java
@@ -282,7 +282,7 @@ public class PopupWindow extends Window implements iPopup {
 
   @Override
   protected void createTarget() {
-    String name = "_popup_window_" + Integer.toHexString(hashCode());
+    String name = "_popup_window_" + Integer.toHexString((int)hashCode());
 
     target = new WindowTarget(Platform.getAppContext(), name, rootPane);
   }
